Home Activity tool bar confirm delete dialog

+++ app/src/main/java/org/tuwaiq/recipes/view/recipeDetails/RecipeDetailsActivity.kt
@@ -1,7 +1,9 @@
 package org.tuwaiq.recipes.view.recipeDetails
 
+import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import android.os.Handler
 import android.widget.Toast
 import androidx.activity.viewModels
 import androidx.cardview.widget.CardView
@@ -14,6 +16,10 @@ import org.tuwaiq.recipes.R
 import org.tuwaiq.recipes.databinding.ActivityRecipeDetailsBinding
 import org.tuwaiq.recipes.model.Recipe
 import org.tuwaiq.recipes.util.Base64Helper
+import cn.pedant.SweetAlert.SweetAlertDialog
+import cn.pedant.SweetAlert.SweetAlertDialog.OnSweetClickListener
+import org.tuwaiq.recipes.view.home.HomeActivity
+
 
 class RecipeDetailsActivity : AppCompatActivity() {
     lateinit var binding: ActivityRecipeDetailsBinding
@@ -28,20 +34,34 @@ class RecipeDetailsActivity : AppCompatActivity() {
 
         var recipe = intent.getSerializableExtra("recipe") as Recipe
 
-        if (currentUser?.uid == recipe.uid){
+        if (currentUser?.uid == recipe.uid) {
             binding.editButton.isVisible = true
             binding.deleteButton.isVisible = true
             binding.deleteButton.setOnClickListener {
-                vm.deleteRecipe(recipe.id).observe(this,{
-                    if (it){
-                        Toast.makeText(this, "Your recipe deleted successfully", Toast.LENGTH_SHORT).show()
-                        finish()
-                    }
-                    else{
-                        Toast.makeText(this, "Failed to delete", Toast.LENGTH_SHORT).show()
-                    }
+                SweetAlertDialog(this, SweetAlertDialog.WARNING_TYPE)
+                    .setTitleText("Are you sure?")
+                    .setContentText("Won't be able to recover this recipe!")
+                    .setConfirmText("Yes,delete it!")
+                    .setConfirmClickListener { sDialog ->
+                        vm.deleteRecipe(recipe.id).observe(this, {
+                            if (it) {
+                                //Toast.makeText(this, "Your recipe deleted successfully", Toast.LENGTH_SHORT).show()
 
-                })
+                                sDialog
+                                    .setTitleText("Deleted!")
+                                    .setContentText("Your recipe has been deleted!")
+                                    .setConfirmText("OK")
+                                    .setConfirmClickListener{
+                                        finish()
+                                    }
+                                    .changeAlertType(SweetAlertDialog.SUCCESS_TYPE)
+                            } else {
+                                Toast.makeText(this, "Failed to delete", Toast.LENGTH_SHORT).show()
+                            }
+
+                        })
+                    }
+                    .show()
             }
         }
 
@@ -50,8 +70,7 @@ class RecipeDetailsActivity : AppCompatActivity() {
             var image = Base64Helper.decodeImage(this, recipe.image)
             binding.imageViewDetails.setImageBitmap(image)
             //Picasso.get().load(image).into(binding.imageViewDetails)
-        }
-        else {
+        } else {
             Picasso.get().load(recipe.image).into(binding.imageViewDetails)
         }
         binding.textViewDetailsName.text = recipe.title
@@ -60,10 +79,9 @@ class RecipeDetailsActivity : AppCompatActivity() {
         binding.expandableIngrTV.text = recipe.ingredients
         cardViewIngr.setOnClickListener {
             expandableIngr.isVisible = !expandableIngr.isVisible
-            if (!expandableIngr.isVisible){
+            if (!expandableIngr.isVisible) {
                 binding.imageView6.setImageResource(R.drawable.arrow_right)
-            }
-            else
+            } else
                 binding.imageView6.setImageResource(R.drawable.down_arrow)
         }
 
@@ -71,11 +89,11 @@ class RecipeDetailsActivity : AppCompatActivity() {
         binding.expandableInstructionsTV.text = recipe.instructions
         binding.cardViewInstruction.setOnClickListener {
             binding.expandablInstruction.isVisible = !binding.expandablInstruction.isVisible
-            if (!binding.expandablInstruction.isVisible){
+            if (!binding.expandablInstruction.isVisible) {
                 binding.imageView7.setImageResource(R.drawable.arrow_right)
-            }
-            else
-                binding.imageView7.setImageResource(R.drawable.down_arrow)        }
+            } else
+                binding.imageView7.setImageResource(R.drawable.down_arrow)
+        }
         setContentView(binding.root)
     }
 }
